==> 1_script/4A_RunGAMMbootstrap_PeentDetectionsXLatTSSSXLatOrdinalXLat.R <==
library(boot)
library(dplyr)
library(gam)
library(gamm4)
library(ggplot2)
library(grid)
library(gridExtra)
library(jpeg)
library(lme4)
library(lubridate)
library(mgcv)
library(MuMIn)
library(plotly)
library(pROC)
library(tidyr)
library(unmarked)
library(viridis)


my.theme <- theme_classic() +
  theme(text=element_text(size=16, family="Arial"),
        axis.text.x=element_text(size=16),
        axis.text.y=element_text(size=8),
        axis.title.x=element_text(margin=margin(16,0,0,0)),
        axis.title.y=element_text(margin=margin(0,16,0,0)),
        axis.line.x=element_line(linetype=1),
        axis.line.y=element_line(linetype=1))

# function to obtain random samples stratified by site 
#cannot stratify by 2-week periods because after excluding sites
#without temperature readings one site (13577)
#only had samples from 2 nights within 1 2-week period
fSampleNint <- function(dat, intervals, n) {
  intervals <- enquo(intervals)
  dat %>%
    group_by(sm_id)%>%
    filter(UQ(intervals) %in% sample(unique(UQ(intervals)), n)) %>%
    slice(sample(row_number()))
}#dropped season because after removing nights without temperature
#at least one site only had 2 nights available.
#I could add twilight period to ensure even distribution of samples
#throughout the night. If I do so, then 6 sites will have fewer
#samples because they lack intervals from "Night" and "Astronomical"
#periods on nights with temperature data. All 6 sites
#are northern sites

#Check how number of peents counted per interval varies
peentsint<-read.csv("0_data/processed/3_PeentsMapped_SunAndMoon/TempSunMoonPeentDetections.10min.int.csv", header=TRUE)

LATS<-unique(peentsint$latitude)
LONGS<-unique(peentsint$longitude)
cor(LATS,LONGS)#-0.96


hist(LATS)
library(BAMMtools)
#2 groups
getJenksBreaks(LATS, 1, subset = NULL)
#45.35199: 1 below, 22 above

#3 groups
getJenksBreaks(LATS, 2, subset = NULL)
#45.35199 62.73417

#4 groups
getJenksBreaks(LATS, 3, subset = NULL)
#45.35199 54.32880 62.73417

getJenksBreaks(LATS, 4, subset = NULL)
#45.35199 48.25271 56.73500 62.73417

peentsint$PEENTS<-ifelse(is.na(peentsint$eventID),0,1)
peentsint$intervals<-as.factor(peentsint$intervals)
peentsint$latitude.F<-ifelse(peentsint$latitude<50,"A",
                             ifelse(peentsint$latitude<55,"B",
                                    ifelse(peentsint$latitude<60,"C","D")))
peentsint$halfmonth<-ifelse(peentsint$day<16,"1","2")
peentsint$season<-as.factor(paste0(peentsint$month,"_",peentsint$halfmonth))

peentsint$start_time_new<- ymd_hms(as.character(peentsint$start_time_new))
peentsint$julian<-yday(as.Date(peentsint$date))
peentsint$hour<-hour(peentsint$start_time_new)
peentsint$minute<-minute(peentsint$start_time_new)
peentsint$start_time_numeric<-peentsint$hour+(peentsint$minute/60)

peentsint$TSSS.s<-scale(peentsint$TSSScorr, center=FALSE, scale=TRUE)#scaled to reduce rank deficiency of GAMMs
#use TSSScorr as there were inaccuracies in TSSS
peentsint$ordinal.s<-scale(as.vector(peentsint$julian), center=FALSE, scale=TRUE)#scaled to reduce rank deficiency of GAMMs
peentsint$latitude.s<-scale(peentsint$latitude, center=FALSE, scale=TRUE)#scaled to reduce rank deficiency of GAMMs
peentsint$longitude.s<-scale(peentsint$longitude, center=FALSE, scale=TRUE)#scaled to reduce rank deficiency of GAMMs
peentsint$altitude.s<-scale(peentsint$altitude, center=FALSE, scale=TRUE)#scaled to reduce rank deficiency of GAMMs
peentsint$meantemp.s<-scale(peentsint$meantemp, center=FALSE, scale=TRUE)#not currently being used in these GAMMs
#peentsint$moon.s<-scale(peentsint$moon.fraction, center=FALSE, scale=TRUE)#not currently being used in these GAMMs


peentsint$sm_id<-as.factor(peentsint$sm_id)
peentsint$intervals<-as.factor(peentsint$intervals)
peentsintB<-peentsint[!peentsint$month==8,]#drop August
peentsintC<-peentsintB[!is.na(peentsintB$meantemp),]#drop intervals missing temperature peentsint
maxdur<-max(peentsintC$duration)
peentsintD<-peentsintC[peentsintC$duration==maxdur,]#gets rid of any "remainder" intervals

#this is the point where I could split into training and testing data for purposes of validation
#training data is what I draw samples from
#the test data is then summarized separately afterwards
#and predictions are generated by creating a test data 
#model matrix and matrix-multiplying it with each column of
#bootstrapped coefficients to get 1 column of predictions
#I then graph the observed:predicted relationship and get
#R-squared or correlation coefficient

#I then get a column of predicted values for each bootstrap that I can compare to actual numbers counted
nrow(peentsintD)#75157
peentsintD$sm_id.interval<-as.factor(paste0(peentsintD$sm_id,peentsintD$intervals))
peentsintD.test <- fSampleNint(peentsintD, intervals, 20)
nrow(peentsintD.test)#1886
write.csv(peentsintD.test, file="0_data/processed/4A_PeentActivityRates_GAMMs/peentsintD.test.csv")
peentsintD.train  <- peentsintD[!peentsintD$sm_id.interval %in% peentsintD.test$sm_id.interval,]
nrow(peentsintD.train)#73271
write.csv(peentsintD.train, file="0_data/processed/4A_PeentActivityRates_GAMMs/peentsintD.train.csv")
#GAMM run once before bootstrap to get model matrix

peentsintD.train<-read.csv("0_data/processed/4A_PeentActivityRates_GAMMs/peentsintD.train.csv", header=TRUE)
#some quick data exploration

#Sites in group A
peentsintD.trainA<-peentsintD.train[peentsintD.train$latitude.F=="A",]
#all 10-minute sample intervals in training data from 45-50 d. N.
nrow(peentsintD.trainA)#12073
peentsintD.trainAdetections<-peentsintD.trainA[!is.na(peentsintD.trainA$eventID),]
nrow(peentsintD.trainAdetections)#7103
hist(peentsintD.trainAdetections$TSSScorr)
#shows number of detections vs. TSSS (peaks after sunset [secondary] and before sunrise [primary])
hist(peentsintD.trainAdetections$julian)
#shows number of detections vs. ordinal day (peaks at start of season)

#Sites in group B
peentsintD.trainB<-peentsintD.train[peentsintD.train$latitude.F=="B",]
#all 10-minute sample intervals in training data from 50-55 d. N.
nrow(peentsintD.trainB)#52099
peentsintD.trainBdetections<-peentsintD.trainB[!is.na(peentsintD.trainB$eventID),]
nrow(peentsintD.trainBdetections)#44840
hist(peentsintD.trainBdetections$TSSScorr)
#shows number of detections vs. TSSS (peaks after sunset [secondary] and before sunrise [primary])
hist(peentsintD.trainBdetections$julian)
#shows number of detections vs. ordinal day (peaks around day 170, second peak around day 195-200)

#Sites in group C
peentsintD.trainC<-peentsintD.train[peentsintD.train$latitude.F=="C",]
#all 10-minute sample intervals in training data from 55-60 d. N.
nrow(peentsintD.trainC)#1180
peentsintD.trainCdetections<-peentsintD.trainC[!is.na(peentsintD.trainC$eventID),]
nrow(peentsintD.trainCdetections)#480
hist(peentsintD.trainCdetections$TSSScorr)
#shows number of detections vs. TSSS (peaks well after sunset and sunrise)
hist(peentsintD.trainCdetections$julian)
#shows number of detections vs. ordinal day (peaks from days 163 to 175)

#Sites in group D
peentsintD.trainD<-peentsintD.train[peentsintD.train$latitude.F=="D",]
#all 10-minute sample intervals in training data from 60-65 d. N.
nrow(peentsintD.trainC)#1180
peentsintD.trainDdetections<-peentsintD.trainD[!is.na(peentsintD.trainD$eventID),]
nrow(peentsintD.trainDdetections)#6112
hist(peentsintD.trainDdetections$TSSScorr)
#shows number of detections vs. TSSS (peaks around/before sunset and declines over time)
hist(peentsintD.trainDdetections$julian)
#shows number of detections vs. ordinal day (peaks arouns 170 then declines slowly)

peentsintD.train$nightlength<-as.POSIXct(peentsintD.train$SR.time.posix)-as.POSIXct(peentsintD.train$SS.time.posix)
peentsintD.train$TSSS.SR<-peentsintD.train$nightlength*3600

d<-fSampleNint(peentsintD.train, intervals, 20)

peents.pervisit<-d %>%
  group_by(sm_id, intervalsP) %>% 
  summarize(numdetect= sum(PEENTS), 
            duration=mean(duration_new),
            TSSS=mean(TSSScorr), 
            TSSS.s=mean(TSSS.s), 
            meantemp=mean(meantemp),
            meantemp.s=mean(meantemp.s),
            twilightperiod=names(which.max(table(twilightperiod))),
            latitude.F=names(which.max(table(latitude.F))),
            latitude=mean(latitude),
            latitude.s=mean(latitude.s),
            longitude=mean(longitude),
            longitude.s=mean(longitude.s),
            altitude=mean(altitude),
            altitude.s=mean(altitude.s),
            #moon.fraction=mean(fraction),
            date=names(which.max(table(date))),
            ordinal.day=mean(julian),
            ordinal.s=mean(ordinal.s),
            time=mean(start_time_numeric))
write.csv(peents.pervisit, file="0_data/processed/4A_PeentActivityRates_GAMMs/temp/peentspervisit.csv")
write.csv(peentsintD.train, file="0_data/processed/4A_PeentActivityRates_GAMMs/temp/peentspervisittraining.csv")
#confirms that after summarizing I have 10 sample observations per site
#at this point, I can run mixed-effects models or GAMMs

#GAMM

#basic model for 10-minute peent intervals: run once to get model matrix for making predictions
peents.pervisit$latitude.F<-as.factor(peents.pervisit$latitude.F)

mean(peents.pervisit$numdetect)  
var(peents.pervisit$numdetect)  
thetaval=(var(peents.pervisit$numdetect)/mean(peents.pervisit$numdetect))

#full.GAM.SS <- gamm4(numdetect ~latgroupN + s(meantemp.s, bs = "cs", by=latgroupN) + s(ordinal.s, bs = "cs", by=latgroupN) + s(TSSS.s, bs = "cs", by=latgroupN), random=~(1|sm_id), data=peents.pervisit, family=negbin(theta=thetaval, link="log"))
#full.GAM.SS <- gamm4(numdetect ~latitude.F + s(meantemp.s, k = 4, bs = "cs", by=latitude.F) + s(ordinal.s, k = 4, bs = "cs", by=latitude.F) + s(TSSS.s, k = 4, bs = "cs", by=latitude.F), random=~(1|sm_id), data=peents.pervisit, family=negbin(theta=thetaval, link="log"))
#plot.gam(full.GAM.SS$gam, main = "Smoothed Untransformed Boom Activity Estimates")
#coef.gamm<-coef(full.GAM.SS$gam)#40 coefficients
full.GAM.SS <- gamm4(numdetect ~latitude.F + s(ordinal.s, k = 3, bs = "cs", by=latitude.F) + s(TSSS.s, k = 5, bs = "cs", by=latitude.F), random=~(1|sm_id), data=peents.pervisit, family=negbin(theta=thetaval, link="log"))
#Changed number of knots from 4 and 4 to 3 and 5
coef.gammB<-coef(full.GAM.SS$gam)
model.matrix(full.GAM.SS$gam)
# "(Intercept)","latitude.FB","latitude.FC","latitude.FD", 
# "s(ordinal.s):latitude.FA.1","s(ordinal.s):latitude.FA.2","s(ordinal.s):latitude.FB.1","s(ordinal.s):latitude.FB.2", 
# "s(ordinal.s):latitude.FC.1","s(ordinal.s):latitude.FC.2","s(ordinal.s):latitude.FD.1","s(ordinal.s):latitude.FD.2", 
# "s(TSSS.s):latitude.FA.1","s(TSSS.s):latitude.FA.2","s(TSSS.s):latitude.FA.3","s(TSSS.s):latitude.FA.4", 
# "s(TSSS.s):latitude.FB.1","s(TSSS.s):latitude.FB.2","s(TSSS.s):latitude.FB.3","s(TSSS.s):latitude.FB.4", 
# "s(TSSS.s):latitude.FC.1","s(TSSS.s):latitude.FC.2","s(TSSS.s):latitude.FC.3","s(TSSS.s):latitude.FC.4", 
# "s(TSSS.s):latitude.FD.1","s(TSSS.s):latitude.FD.2","s(TSSS.s):latitude.FD.3","s(TSSS.s):latitude.FD.4"

df<-data.frame(coef.gamm)#28 coefficients
#write.csv(df, file="3_output/data/4A_PeentGAMMS/GAMMpeents.10min.int.csv")

#Now create the bootstrap
bs <- function(data, indices){
  d<-fSampleNint(data, intervals, 20)
  
  peents.pervisit<-d %>%
    group_by(sm_id, intervalsP) %>% 
    summarize(numdetect= sum(PEENTS), 
              duration=mean(duration_new),
              TSSS=mean(TSSScorr), 
              TSSS.s=mean(TSSS.s), 
              meantemp=mean(meantemp),
              meantemp.s=mean(meantemp.s),
              twilightperiod=names(which.max(table(twilightperiod))),
              latitude=mean(latitude),
              latitude.F=names(which.max(table(latitude.F))),
              latitude.s=mean(latitude.s),
              longitude=mean(longitude),
              longitude.s=mean(longitude.s),
              altitude=mean(altitude),
              altitude.s=mean(altitude.s),
              #moon.fraction=mean(fraction),
              date=names(which.max(table(date))),
              ordinal.day=mean(julian),
              ordinal.s=mean(ordinal.s),
              time=mean(start_time_numeric))
  
  #GAMM

  #basic model for 10-minute peent intervals
  peents.pervisit$latitude.F<-as.factor(peents.pervisit$latitude.F)
  mean(peents.pervisit$numdetect)  
  var(peents.pervisit$numdetect)  
  thetaval=(var(peents.pervisit$numdetect)/mean(peents.pervisit$numdetect))
  
  try(full.GAM.SS <- gamm4(numdetect ~latitude.F  + s(ordinal.s, k = 3, bs = "cs", by=latitude.F) + s(TSSS.s, k = 5, bs = "cs", by=latitude.F), random=~(1|sm_id), data=peents.pervisit, family=negbin(theta=thetaval, link="log")))
  #plot.gam(full.GAM.SS$gam, main = "Smoothed Untransformed Peent Activity Estimates")
  try(coef.gamm<-coef(full.GAM.SS$gam))
  try(return(coef.gamm))
}

#peent.all<-peentsintD.train#read.csv("0_data/processed/3_PeentsMapped_SunAndMoon/TempSunMoonPeentDetections.10min.int.csv", header=TRUE)

peentresults <- boot(data=peentsintD.train,
                    statistic=bs, parallel="multicore",
                    R=100)
#print(paste0("bootstraps successfully run for ",fn))

CoefB <-t(peentresults$t)
#get rid of columns with errors
CoefB<-data.frame(CoefB)
CoefB[] <- lapply(CoefB, function(x) as.numeric(as.character(x)))
#delete columns with NA values
all_na <- function(x) any(!is.na(x))
CoefB<-CoefB %>% select_if(all_na)
CoefB<-as.matrix(CoefB)
ncol(CoefB)
CoefBCI50 <- t(apply(CoefB, 1, quantile, c(0.5, 0.25, 0.75), na.rm=TRUE))
CoefBCI90 <- t(apply(CoefB, 1, quantile, c(0.5, 0.05, 0.95), na.rm=TRUE))
save(peentresults, CoefB, CoefBCI50, CoefBCI90, file="3_output/data/4A_March2021_BootstrappedPeentActivityGAMMs/boot20samples10minuteintPeents.RData")
#save as csv
df<-data.frame(CoefB)
row.names(df) = c("(Intercept)","latitude.FB","latitude.FC","latitude.FD", 
                  "s(ordinal.s):latitude.FA.1","s(ordinal.s):latitude.FA.2","s(ordinal.s):latitude.FB.1","s(ordinal.s):latitude.FB.2", 
                  "s(ordinal.s):latitude.FC.1","s(ordinal.s):latitude.FC.2","s(ordinal.s):latitude.FD.1","s(ordinal.s):latitude.FD.2", 
                  "s(TSSS.s):latitude.FA.1","s(TSSS.s):latitude.FA.2","s(TSSS.s):latitude.FA.3","s(TSSS.s):latitude.FA.4", 
                  "s(TSSS.s):latitude.FB.1","s(TSSS.s):latitude.FB.2","s(TSSS.s):latitude.FB.3","s(TSSS.s):latitude.FB.4", 
                  "s(TSSS.s):latitude.FC.1","s(TSSS.s):latitude.FC.2","s(TSSS.s):latitude.FC.3","s(TSSS.s):latitude.FC.4", 
                  "s(TSSS.s):latitude.FD.1","s(TSSS.s):latitude.FD.2","s(TSSS.s):latitude.FD.3","s(TSSS.s):latitude.FD.4")
write.csv(df, file="3_output/data/4A_March2021_BootstrappedPeentActivityGAMMs/boot20samples10minuteintPeents.csv")
#print(paste0("bootstrap coefficients saved for ",fn))

## box plot of variables with confidence intervals
prednames = c("(Intercept)","latitude.FB","latitude.FC","latitude.FD", 
              "s(ordinal.s):latitude.FA.1","s(ordinal.s):latitude.FA.2","s(ordinal.s):latitude.FB.1","s(ordinal.s):latitude.FB.2", 
              "s(ordinal.s):latitude.FC.1","s(ordinal.s):latitude.FC.2","s(ordinal.s):latitude.FD.1","s(ordinal.s):latitude.FD.2", 
              "s(TSSS.s):latitude.FA.1","s(TSSS.s):latitude.FA.2","s(TSSS.s):latitude.FA.3","s(TSSS.s):latitude.FA.4", 
              "s(TSSS.s):latitude.FB.1","s(TSSS.s):latitude.FB.2","s(TSSS.s):latitude.FB.3","s(TSSS.s):latitude.FB.4", 
              "s(TSSS.s):latitude.FC.1","s(TSSS.s):latitude.FC.2","s(TSSS.s):latitude.FC.3","s(TSSS.s):latitude.FC.4", 
              "s(TSSS.s):latitude.FD.1","s(TSSS.s):latitude.FD.2","s(TSSS.s):latitude.FD.3","s(TSSS.s):latitude.FD.4")
var.bci<-CoefBCI90
var.bci<-data.frame(var.bci)
var.bci$median<-as.numeric(var.bci$X50.)
var.bci$lcl<-as.numeric(var.bci$X5.)
var.bci$ucl<-as.numeric(var.bci$X95.)
var.bci$prednames<-prednames

GG<-ggplot(var.bci, aes(x=prednames, y=median))+
  geom_point(aes(x=prednames, y=median))+
  geom_errorbar(aes(ymin=lcl,ymax=ucl))+
  geom_hline(yintercept=0)+
  xlab("Predictor")+
  ylab("Effect on peents counted")+coord_flip()+my.theme
ggsave("3_output/figures/4A_March2021_BootstrappedPeentActivityGAMMs/boot20visits10minuteintervalPeent.jpeg", plot=GG, width=12, height=6, units=c("in"), dpi=300)

#To get GAMM predictions within a bootstrap:
#https://gist.github.com/davharris/3da3a43f3d45ce01ee20

## make data for prediction - TSSS
load("3_output/data/4A_March2021_BootstrappedPeentActivityGAMMs/boot20samples10minuteintPeents.RData")
#gives us CoefB

library(ggplot2)
library(viridis)

#Creating datasets, then matrix plots of predicted activity (one curve per latitude and date)
peents<-read.csv("0_data/processed/4A_PeentActivityRates_GAMMs/temp/peentspervisittraining.csv", header=TRUE)#peentspervisittraining.csv
dates<-unique(peents[,c("date","julian","ordinal.s")])
dates<-data.frame(dates)
#gives us a list of dates, ordinal days, and scaled values
#to loop through

sunrisetimes<-peents%>%
  group_by(date,latitude.F)%>%
  summarise(TSSS.SR=mean(TSSS.SR))
sunrisetimes<-data.frame(sunrisetimes)
sunrisetimes$latitude.F<-as.factor(sunrisetimes$latitude.F)
#gives us a list of mean sunrise times on different dates for
#sites in different latitude groups


range(peents$julian)#152 212
range(peents$TSSScorr)#-3600 36600

#Matrix plots: Peents X TSSS on different days at different latitudes
for (i in 1:nrow(dates)){
  Day<-dates[i,1]
  ordinal.day<-dates[i,2]
  peent.tsss.jd<-expand.grid(intercept=1,
                             latitude.F=c("A","B","C","D"), 
                             TSSS=seq(from=-3600,to=36600,by=(36600--3600)/100),
                             OrdinalDay=seq(from=152,to=212,by=2)
  )
  peent.tsss.jd$latitude.FB<-ifelse(peent.tsss.jd$latitude.F=="B",1,0)
  peent.tsss.jd$latitude.FC<-ifelse(peent.tsss.jd$latitude.F=="C",1,0)
  peent.tsss.jd$latitude.FD<-ifelse(peent.tsss.jd$latitude.F=="D",1,0)
  Latitude<-peent.tsss.jd$latitude.F
  #peent.tsss.jd$ordinal.day<-dates[i,2]#
  peent.tsss.jd$ordinal.s<-scale(peent.tsss.jd$OrdinalDay, center=FALSE, scale=TRUE)#scaled value associated with ordinal day
  peent.tsss.jd$TSSS.s<-scale(peent.tsss.jd$TSSS, center=FALSE, scale=TRUE)#scaled to reduce rank deficiency of GAMMs
  #peent.tsss.jd$LxO<-peent.tsss.jd$latitude.s*peent.tsss.jd$Ordinal.orth1#peent.tsss.jd$sm_id<-"7524"
  #peent.tsss.jd$LxT<-peent.tsss.jd$latitude.s*peent.tsss.jd$TSSS.orth1#peent.tsss.jd$sm_id<-"7524"
  #peent.tsss.jd$LxT2<-peent.tsss.jd$latitude.s*peent.tsss.jd$TSSS.orth2#peent.tsss.jd$sm_id<-"7524"
  #peent.tsss.jd$LxO2<-peent.tsss.jd$latitude.s*peent.tsss.jd$Ordinal.orth2#peent.tsss.jd$sm_id<-"7524"
  #peent.tsss.jd$sm_id<-as.factor(peent.tsss.jd$sm_id)
  
  #Now limit predictions to those from a single date
  peent.tsss.jd<-peent.tsss.jd[peent.tsss.jd$OrdinalDay==ordinal.day,]
  
  #set unscaled latitude and TSSS aside for now and remove these
  #variables from prediction dataset prior to creating fixed effects
  #design matrix
  OrdinalDay<-peent.tsss.jd$OrdinalDay
  #ORDINAL.S<-peent.tsss.jd$ordinal.s
  #peent.tsss.jd$OrdinalDay<-NULL
  #peent.tsss.jd$ordinal.s<-NULL
  #peent.tsss.jd$ordinal.day<-NULL
  Latitude<-peent.tsss.jd$latitude.F
  #peent.tsss.jd$latitude.F<-NULL
  TSSS<-peent.tsss.jd$TSSS
  #peent.tsss.jd$TSSS<-NULL
  #peent.tsss.jd$TSSS.s<-NULL
  str(peent.tsss.jd)
  names(peent.tsss.jd)
  
  #peent.tsss.jdmm<-as.matrix(peent.tsss.jd)
  peent.tsss.jdmm <- model.matrix(full.GAM.SS$gam, peent.tsss.jd)#meantemp.s+moon.s+
  
  ## predict and apply inverse link function: this gives an n_new x B+1 matrix
  bootpreds100 <- (exp(peent.tsss.jdmm %*% CoefB))#take antilog?
  
  ## calculate median and 95% CIs for predictions
  bp100.CI95 <- t(apply(bootpreds100, 1, quantile, seq(from=0.025, to=0.975, by=((0.975-0.025)/95)), na.rm=TRUE))
  
  ## Use these bootstrapped predictions to generate heat maps, 3d plots, etc.
  
  peent.tsss.jd.bci<-cbind(peent.tsss.jd, bp100.CI95)
  peent.tsss.jd.F<-cbind(peent.tsss.jd.bci,Latitude,TSSS,OrdinalDay)#put Latitude and TSSS back in
  peent.tsss.jd.F$TSSS.min<-peent.tsss.jd.F$TSSS/60
  peent.tsss.jd.F$TSSS.hr<-peent.tsss.jd.F$TSSS/3600
  peent.tsss.jd.F$Latitude.F<-as.factor(peent.tsss.jd.F$Latitude)
  
  levels(peent.tsss.jd.F$Latitude.F)
  latlevels<-c("A","B","C","D")
  for (j in 1:length(latlevels)){
    preds.sitelevel<-peent.tsss.jd.F[peent.tsss.jd.F$Latitude.F==latlevels[j],]
    preds.025.975<-preds.sitelevel[,10:96]
    for (k in 1:ncol(preds.025.975)){
      preds.025.975[,k]<-ifelse(preds.025.975[,k]>600,600,preds.025.975[,k])
    }
    sunrisetimesB<-sunrisetimes[sunrisetimes$latitude.F==latlevels[j],]
    sunrisetimesC<-sunrisetimesB[sunrisetimesB$date==Day,]
    if (nrow(sunrisetimesC)>0){
      preds.sitelevel$TSSS.SR<-sunrisetimesC$TSSS.SR
      preds.sitelevel$TSSS.SR.hr<-preds.sitelevel$TSSS.SR/3600
    }
    if (nrow(sunrisetimesC)==0){
      preds.sitelevel$TSSS.SR.hr<-0#if no sunrise times are available for that combination of date and latitude
    }
    tiff(paste0("3_output/figures/4A_BootstrappedPeentGAMMs/Matrix Plots peents X TSSS X lat different days March 17/",ifelse(j==1,"A",ifelse(j==2,"B_",ifelse(j==3,"C","D"))),"PeentvsTSSS-",preds.sitelevel$Latitude.F,"-",preds.sitelevel$OrdinalDay,".tiff"), units="in", width=6, height=4, res=300)
    X <- c(0,unique(preds.sitelevel$TSSS.SR.hr))
    Y <- c(0,max(preds.025.975))
    matplot(preds.sitelevel$TSSS.hr, 
            preds.025.975, 
            type="l", 
            col="blue",#cividis(100),
            xlab=paste0("Hours After Sunset ", Day),
            #xlim=rev(range(preds.sitelevel$TSSS.hr)),
            ylab=paste0("Peent Activity (",ifelse(j==1,"45-50 deg. N ", ifelse(j==2,"50-55 deg. N",ifelse(j==3,"55-60 deg. N","60-65 deg. N"))),")"))   
    abline(v=c(0,unique(preds.sitelevel$TSSS.SR.hr)), col=c("red", "red"), lty=c(1,2), lwd=c(1, 3))
    #lim <- par("usr")
    #rect(X[1], Y[1], X[2], Y[2], border = "red", col = NA)
    #axis(1) ## add axes back
    #axis(2)
    #box()
    dev.off()
  }
}

#Get multipanel plot of peents vs. TSSS
library(magick)
library(dplyr)
library(tidyr)
library(magrittr)
# read images and then create a montage
# tile =2 , means arrange the images in 2 columns
# geometry controls the pixel sixe, spacing between each image in the collage output. 
# read the the png files into a list
pngfiles <-
  list.files(
    path = "3_output/figures/4A_BootstrappedPeentGAMMs/Matrix Plots peents X TSSS X lat different days March 17",
    recursive = TRUE,
    pattern = "\\.png$",
    full.names = T
  )

magick::image_read(pngfiles) %>%
  magick::image_montage(tile = "3", geometry = "x500+10+5") %>%
  magick::image_convert("jpg") %>%
  magick::image_write(
    format = ".png", path = "3_output/figures/4A_BootstrappedPeentGAMMs/Figure4_peentsXtsss_collage.png",
    quality = 100
  )

#Matrix plots: Peents X different days 1 hour after sunset at different latitudes
  peent.tsss.jd<-expand.grid(intercept=1,
                             latitude.F=c("A","B","C","D"), 
                             TSSS=3600,
                             OrdinalDay=seq(from=152,to=212,by=2)
  )
  peent.tsss.jd$latitude.FB<-ifelse(peent.tsss.jd$latitude.F=="B",1,0)
  peent.tsss.jd$latitude.FC<-ifelse(peent.tsss.jd$latitude.F=="C",1,0)
  peent.tsss.jd$latitude.FD<-ifelse(peent.tsss.jd$latitude.F=="D",1,0)
  Latitude<-peent.tsss.jd$latitude.F
  peent.tsss.jd$ordinal.s<-scale(peent.tsss.jd$OrdinalDay, center=FALSE, scale=TRUE)#scaled value associated with ordinal day
  peent.tsss.jd$TSSS.s<-0.216293523#value associated with TSSS=3600 in training data

  #set unscaled latitude and TSSS aside for now and remove these
  #variables from prediction dataset prior to creating fixed effects
  #design matrix
  OrdinalDay<-peent.tsss.jd$OrdinalDay
  Latitude<-peent.tsss.jd$latitude.F
  TSSS<-peent.tsss.jd$TSSS
  str(peent.tsss.jd)
  names(peent.tsss.jd)
  
  #peent.tsss.jdmm<-as.matrix(peent.tsss.jd)
  peent.tsss.jdmm <- model.matrix(full.GAM.SS$gam, peent.tsss.jd)#meantemp.s+moon.s+
  
  ## predict and apply inverse link function: this gives an n_new x B+1 matrix
  bootpreds100 <- (exp(peent.tsss.jdmm %*% CoefB))#take antilog?
  
  ## calculate median and 95% CIs for predictions
  bp100.CI95 <- t(apply(bootpreds100, 1, quantile, seq(from=0.025, to=0.975, by=((0.975-0.025)/95)), na.rm=TRUE))
  
  ## Use these bootstrapped predictions to generate heat maps, 3d plots, etc.
  
  peent.tsss.jd.bci<-cbind(peent.tsss.jd, bp100.CI95)
  peent.tsss.jd.F<-cbind(peent.tsss.jd.bci,Latitude,TSSS,OrdinalDay)#put Latitude and TSSS back in
  peent.tsss.jd.F$TSSS.min<-peent.tsss.jd.F$TSSS/60
  peent.tsss.jd.F$TSSS.hr<-peent.tsss.jd.F$TSSS/3600
  peent.tsss.jd.F$Latitude.F<-as.factor(peent.tsss.jd.F$Latitude)
  
  levels(peent.tsss.jd.F$Latitude.F)
  latlevels<-c("A","B","C","D")
  for (j in 1:length(latlevels)){
    preds.sitelevel<-peent.tsss.jd.F[peent.tsss.jd.F$Latitude.F==latlevels[j],]
    preds.025.975<-preds.sitelevel[,10:96]
    for (k in 1:ncol(preds.025.975)){
      preds.025.975[,k]<-ifelse(preds.025.975[,k]>600,600,preds.025.975[,k])
    }
    tiff(paste0("3_output/figures/4A_BootstrappedPeentGAMMs/Matrix Plots peents X ordinal day TSSS 3600/",ifelse(j==1,"A",ifelse(j==2,"B_",ifelse(j==3,"C","D"))),"PeentvsDate-",preds.sitelevel$Latitude.F,"-",preds.sitelevel$OrdinalDay,".tiff"), units="in", width=6, height=4, res=300)
    Y <- c(0,max(preds.025.975))
    matplot(preds.sitelevel$OrdinalDay, 
            preds.025.975, 
            type="l", 
            col="blue",#cividis(100),
            xlab=paste0("Ordinal Day"),
            ylab=paste0("Peent Activity (",ifelse(j==1,"45-50 deg. N ", ifelse(j==2,"50-55 deg. N",ifelse(j==3,"55-60 deg. N","60-65 deg. N"))),")"))   
    dev.off()
  }

#Get multipanel plot of peents vs. ordinal day
  library(magick)
  library(dplyr)
  library(tidyr)
  library(magrittr)
  # read images and then create a montage
  # tile =2 , means arrange the images in 2 columns
  # geometry controls the pixel sixe, spacing between each image in the collage output. 
  # read the the png files into a list
  pngfiles <-
    list.files(
      path = "3_output/figures/4A_BootstrappedPeentGAMMs/Matrix Plots peents X ordinal day TSSS 3600",
      recursive = TRUE,
      pattern = "\\.png$",
      full.names = T
    )
  
  magick::image_read(pngfiles) %>%
    magick::image_montage(tile = "2", geometry = "x500+10+5") %>%
    magick::image_convert("jpg") %>%
    magick::image_write(
      format = ".png", path = "3_output/figures/4A_BootstrappedPeentGAMMs/Figure6_peentsXordinalday_collage.png",
      quality = 100
    )

#Validating the GAMMS, using the test data from earlier
#i.e. peentsintD.test, which still must be summarized by interval
peentsintD.test<-read.csv("0_data/processed/4A_PeentActivityRates_GAMMs/peentsintD.test.csv") 
peents.pervisit.test<-peentsintD.test %>%
  group_by(sm_id, intervalsP) %>% 
  summarize(numdetect= sum(PEENTS), 
            duration=mean(duration_new),
            TSSS=mean(TSSScorr), 
            TSSS.s=mean(TSSS.s), 
            meantemp=mean(meantemp),
            meantemp.s=mean(meantemp.s),
            twilightperiod=names(which.max(table(twilightperiod))),
            latitude=mean(latitude),
            latitude.s=mean(latitude.s),
            latitude.F=names(which.max(table(latitude.F))),
            longitude=mean(longitude),
            #moon.fraction=mean(fraction),
            date=names(which.max(table(date))),
            ordinal.day=mean(julian),
            ordinal.s=mean(ordinal.s),
            time=mean(start_time_numeric))

peents.pervisit.test$latitude.FB<-ifelse(peents.pervisit.test$latitude.F=="B",1,0)
peents.pervisit.test$latitude.FC<-ifelse(peents.pervisit.test$latitude.F=="C",1,0)
peents.pervisit.test$latitude.FD<-ifelse(peents.pervisit.test$latitude.F=="D",1,0)

## model matrix that'll match the coefficients
Xnew.Test.mm <- model.matrix(full.GAM.SS$gam, peents.pervisit.test)

## predict and apply inverse link function: this gives an n_new x B+1 matrix
Preds <- (exp(Xnew.Test.mm %*% CoefB))

peents.pervisit.test.df<-data.frame(peents.pervisit.test)
Preds.df<-data.frame(Preds)
TestDatapreds<-cbind(peents.pervisit.test.df, Preds.df)

str(TestDatapreds)
write.csv(TestDatapreds, file="3_output/data/4A_March2021_BootstrappedPeentActivityGAMMs/TestDatapreds.csv")

spearmancorlist<-list()
names<-c("X1","X2","X3","X4","X5","X6","X7","X8","X9","X10",
         "X11","X12","X13","X14","X15","X16","X17","X18","X19","X20",
         "X21","X22","X23","X24","X25","X26","X27","X28","X29","X30",
         "X31","X32","X33","X34","X35","X36","X37","X38","X39","X40",
         "X41","X42","X43","X44","X45","X46","X47","X48","X49","X50",
         "X51","X52","X53","X54","X55","X56","X57","X58","X59","X60",
         "X61","X62","X63","X64","X65","X66","X67","X68","X69","X70",
         "X71","X72","X73","X74","X75","X76","X77","X78","X79","X80",
         "X81","X82","X83","X84","X85","X86","X87","X88","X89","X90",
         "X91","X92","X93","X94","X95","X96","X97","X98","X99","X100")
for (i in names){
  TestDatapreds$Predicted<-TestDatapreds[,i]
  spearmancorlist[[i]]<-cor(TestDatapreds$numdetect, TestDatapreds$Predicted, method="spearman")
}
spearmancorvec<-unlist(spearmancorlist)
spearmandf<-data.frame(spearmancorvec)
## calculate median and 90% CIs
RhoN <-quantile(spearmancorvec, c(0.5, 0.05, 0.95), na.rm=TRUE)
#       50%         5%        95% 
#0.13488234 0.02502464 0.22827643 (in 2020)
#0.2202588  0.1002912  0.2777177  (in 2021)
